+++ runner/monitor/java/androidx/test/internal/platform/Supersedes.java
@@ -1,0 +1,17 @@
+package androidx.test.internal.platform;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/** Indicates that the annotated type is intended as a replacement for another type. */
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Supersedes {
+
+  /** The type that is superseded by the annotated type. */
+  Class<?> value();
+}